restful-webservices/src/main/java/com/mmpmr/controller/HelloWorldController.java
package com.mmpmr.controller;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloWorldController {

	@GetMapping(path = "/hello-world")
	public String helloWorld() {
		return "Hello World";
	}

	@GetMapping(path = "/hello/{name}")
	public String helloWithPathVar(@PathVariable String name) {
		return String.format("Hello, %s", name);
	}
}
